api/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
httpx==0.27.0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai==1.42.0
alpaca-trade-api==3.1.1
